integrations/langchain/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "langchain-docdigitizer"
version = "0.1.0"
description = "LangChain document loader for the DocDigitizer document processing API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "DocDigitizer", email = "support@docdigitizer.com" },
]
keywords = ["langchain", "docdigitizer", "document-loader", "ocr", "pdf", "extraction"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "docdigitizer>=0.1.0",
    "langchain-core>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/DocDigitizer/dd-v3-integrations"
Documentation = "https://github.com/DocDigitizer/dd-v3-integrations/tree/main/integrations/langchain/python"
Repository = "https://github.com/DocDigitizer/dd-v3-integrations"

[tool.hatch.build.targets.wheel]
packages = ["src/langchain_docdigitizer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
